pyproject.toml
[project]
name = "statistics-25"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "catboost>=1.2.8",
    "fonttools>=4.60.1",
    "imbalanced-learn>=0.14.0",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pkg>=0.2",
    "pyyaml",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "shap>=0.49.1",
    "tqdm>=4.67.1",
    "xgboost>=3.0.5",
    "PyYAML",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
